+++ hootsweet/constants.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class Reviewer(Enum):
+    EXTERNAL = 1
+    MEMBER = 2